refactor(codex-auth): extract shared Codex auth core out of the import service
resolveCodexUnixIdentity, persistVerifiedCodexAuth, AppLike, and
CodexUnixIdentityResolution were exported from codex-auth-import.ts and imported
by device/logout/check-auth, making a named endpoint double as the shared core.
Move them to codex-auth-shared.ts; codex-auth-import.ts is now just the import
endpoint and consumes the shared module like its siblings. Pure move — no
behavior change.
